use a different recursion
this was very tricky. We notice that there are a lot of calculations we can *reuse*.
The replacement for a given pair of characters after N generations is always the same,
and we need to evaluate the eventual replacement for a given pair of letters after a
given period many times on the way to building the final result.

we can use memoist to take over the repetitive calculations, we just need to write
it as an appropriate recursion. We could actually use the same trick to construct the real
strings, but the count-hashes are a more compact format, only keeping the data we actually
need